confirm delete on the scopes dialog

+++ src/main/webapp/js/views/ClientScopesModal.js
@@ -41,14 +41,66 @@ define([ "react", "util", "underscore", "./Loading", "rbs/components/collection/
         component: "checkbox"
       },
       {
-        component: btn,
-        type: "danger",
-        icon: "ban",
-        size: "xs",
-        ajax: true,
-        onClick: function () {
-          this.props.model.destroy({ wait: true });
-        }
+        component: util.rf({
+          displayName: "delete button",
+          getInitialState: function () {
+            return {
+              confirm: false
+            };
+          },
+
+          cancel: function () {
+            this.setState({
+              confirm: false
+            });
+          },
+          confirm: function () {
+            this.setState({
+              confirm: true
+            });
+          },
+
+          getButtons: function () {
+            if (this.state.confirm) {
+              return [
+                btn({
+                  key: "cancel",
+                  caption: "Cancel",
+                  size: "xs",
+                  ajax: true,
+                  onClick: this.cancel
+                }),
+                btn({
+                  key: "btn",
+                  type: "danger",
+                  caption: "Confirm",
+                  icon: "ban",
+                  size: "xs",
+                  ajax: true,
+                  onClick: _.bind(function () {
+                    this.props.model.destroy({ wait: true });
+                  }, this)
+                })
+              ];
+            } else {
+              return [
+                btn({
+                  key: "btn",
+                  type: "danger",
+                  caption: "Delete",
+                  icon: "ban",
+                  size: "xs",
+                  ajax: true,
+                  onClick: this.confirm
+                })
+              ];
+            }
+          },
+
+          render: function () {
+            return d.div({ className: "text-center" }, d.div({ className: "btn-group" }, this.getButtons()));
+          }
+        })
       }
     ];
